Extracted stateless widgets to simplify code
