+++ lib/widgets/tokens/token_widgets.dart
@@ -50,7 +50,8 @@ import 'package:privacyidea_authenticator/utils/push_provider.dart';
 import 'package:privacyidea_authenticator/utils/storage_utils.dart';
 import 'package:privacyidea_authenticator/utils/utils.dart';
 
-import 'custom_texts.dart';
+import '../custom_texts.dart';
+import 'components.dart';
 
 class TokenWidget extends StatefulWidget {
   final Token _token;
@@ -79,7 +80,7 @@ class TokenWidget extends StatefulWidget {
 
 abstract class _TokenWidgetState extends State<TokenWidget> {
   Token _token;
-  static final SlidableController _slidableController = SlidableController();
+  final SlidableController _slidableController = SlidableController();
 
   _TokenWidgetState(this._token) {
     _saveThisToken();
@@ -97,7 +98,7 @@ abstract class _TokenWidgetState extends State<TokenWidget> {
             ? Colors.black
             : Colors.white,
         icon: Icons.delete,
-        onTap: () => _deleteTokenDialog(),
+        onTap: () => _showDeleteTokenDialog(),
       ),
       IconSlideAction(
         caption: AppLocalizations.of(context)!.customize,
@@ -243,7 +244,7 @@ abstract class _TokenWidgetState extends State<TokenWidget> {
     return didAuthenticate;
   }
 
-  void _deleteTokenDialog() {
+  void _showDeleteTokenDialog() {
     showDialog(
         context: context,
         builder: (BuildContext context) {
@@ -967,78 +968,6 @@ class _TotpWidgetState extends _OTPTokenWidgetState
         LinearProgressIndicator(
           value: _controller.value,
         ),
-      ],
-    );
-  }
-}
-
-class TokenTile extends StatelessWidget {
-  final Widget _avatar;
-  final Widget _title;
-  final Widget _subtitle;
-
-  TokenTile({avatar, title, subtitle})
-      : _avatar = avatar,
-        _title = title,
-        _subtitle = subtitle;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Padding(
-          padding: EdgeInsets.all(5),
-          child: _avatar,
-        ),
-        Expanded(
-          child: ListTile(
-            title: _title,
-            subtitle: _subtitle,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class TokenAvatar extends StatelessWidget {
-  final int? _avatarColor;
-  final String? _avatarImagePath;
-  final double _radius;
-
-  const TokenAvatar(this._avatarColor, this._avatarImagePath, this._radius);
-
-  @override
-  Widget build(BuildContext context) {
-    return CircleAvatar(
-      backgroundColor: _avatarColor == null ? null : Color(_avatarColor!),
-      backgroundImage:
-          _avatarImagePath == null ? null : FileImage(File(_avatarImagePath!)),
-      radius: _radius,
-    );
-  }
-}
-
-class TokenSubtitle extends StatelessWidget {
-  final Token _token;
-
-  const TokenSubtitle(this._token);
-
-  @override
-  Widget build(BuildContext context) {
-    TextStyle style = Theme.of(context)
-        .textTheme
-        .headline6!
-        .copyWith(fontWeight: FontWeight.normal);
-
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        _token is PushToken
-            ? Text((_token as PushToken).serial, style: style)
-            : Text(_token.label, style: style),
-        Text(_token.issuer, style: style)
       ],
     );
   }
